+++ cdl/core/gui/processor/signal.py
@@ -286,6 +286,16 @@ class SignalProcessor(BaseProcessor):
         )
 
     @qt_try_except()
+    def compute_convolution(self, obj2: SignalObj | None = None) -> None:
+        """Compute convolution"""
+        self.compute_n1n(
+            obj2,
+            _("signal to convolve with"),
+            cps.compute_convolution,
+            title=_("Convolution"),
+        )
+
+    @qt_try_except()
     def compute_fit(self, name, fitdlgfunc):
         """Compute fitting curve"""
         for obj in self.panel.objview.get_sel_objects():